Changed doXYPixelOffset in ACQUIREImplementation.java / ACQUIREImplementationTweak.java so it takes doubles rather than ints. This stops pixel quantisation of the OFFSET_X_Ys improving the acquisition (in theory).

+++ java/ngat/sprat/ACQUIREImplementation.java
@@ -487,8 +487,8 @@ public class ACQUIREImplementation extends FITSImplementation implements JMSComm
 	 * The plate scale is currently retrieved from the FITS property file value:"sprat.fits.value.CCDSCALE."+bin.
 	 * It is not clear how non-square binning is supported yet, or whether this is the right thing to do.
 	 * @param id The string id of the command instance we are implementing. Used for generating ISS command id's.
-	 * @param xPixelOffset The offset in X binned pixels in the focal plane of the acquisition instrument.
-	 * @param yPixelOffset The offset in Y binned pixels in the focal plane of the acquisition instrument.
+	 * @param xPixelOffset The offset in X binned decimal pixels in the focal plane of the acquisition instrument.
+	 * @param yPixelOffset The offset in Y binned decimal pixels in the focal plane of the acquisition instrument.
 	 * @exception Exception Thrown if the ISS command OFFSET_RA_DEC fails.
 	 * @see #sprat
 	 * @see #serverConnectionThread
@@ -498,7 +498,7 @@ public class ACQUIREImplementation extends FITSImplementation implements JMSComm
 	 * @see Sprat#sendISSCommand
 	 * @see ngat.message.ISS_INST.OFFSET_X_Y
 	 */
-	protected void doXYPixelOffset(String id,int xPixelOffset,int yPixelOffset) throws Exception
+	protected void doXYPixelOffset(String id,double xPixelOffset,double yPixelOffset) throws Exception
 	{
 		OFFSET_X_Y offsetXYCommand = null;
 		INST_TO_ISS_DONE instToISSDone = null;
@@ -547,7 +547,8 @@ public class ACQUIREImplementation extends FITSImplementation implements JMSComm
 	 * </ul>
 	 * @param acquireCommand The instance of ACQUIRE we are currently running.
 	 * @param acquireDone The instance of ACQUIRE_DONE to fill in with errors we receive.
-	 * @exception Exception Thrown if testAbort diddly... failed.	 
+	 * @exception Exception Thrown if testAbort/sendBasicAck failed, or if we issued more than
+	 *            maximumOffsetCount offsets.	 
 	 * @see #doFrame
 	 * @see #raOffset
 	 * @see #decOffset
@@ -705,7 +706,7 @@ public class ACQUIREImplementation extends FITSImplementation implements JMSComm
 			if(done == false)
 			{
 				// issue new XY Pixel offset
-				doXYPixelOffset(acquireCommand.getId(),(int)xPixelOffset,(int)yPixelOffset);
+				doXYPixelOffset(acquireCommand.getId(),xPixelOffset,yPixelOffset);
 				offsetCount++;
 			}
 			// Have we taken too many goes to acquire?
